requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.28.2
jupyter>=1.0.0
notebook>=6.5.3 